requirements.txt
# Core dependencies
fastapi>=0.114.0
uvicorn[standard]>=0.30.6
pydantic>=2.8.2
pydantic-settings>=2.4.0
python-dotenv>=1.0.1

# Database
sqlalchemy>=2.0.32
psycopg2-binary>=2.9.9
alembic>=1.13.2

# HTTP and networking
httpx>=0.27.0
aiohttp>=3.9.5
requests>=2.32.3

# Monitoring and observability
prometheus-client>=0.20.0
starlette-exporter>=0.18.0
opentelemetry-sdk>=1.26.0
opentelemetry-exporter-otlp>=1.26.0
opentelemetry-instrumentation-fastapi>=0.47b0
opentelemetry-instrumentation-httpx>=0.47b0
opentelemetry-instrumentation-sqlalchemy>=0.47b0
opentelemetry-instrumentation-redis>=0.47b0

# Caching and messaging
redis>=5.0.7
nats-py>=2.7.2

# AWS integration
boto3>=1.34.159

# Security and cryptography
cryptography>=43.0.1

# Testing
pytest>=8.3.2
pytest-asyncio>=0.23.8

# Utilities
PyYAML>=6.0.2
python-multipart>=0.0.9
websockets>=12.0
croniter>=3.0.3

# ML and AI (for future use)
# torch>=2.0.0
# transformers>=4.30.0
# scikit-learn>=1.3.0
# numpy>=1.24.0
# pandas>=2.0.0
